NormalizeBase in lib-builtin-effects
